real addon location and management
